analytics/GNN_and_GS/requirements.txt
# base
attrdict==2.0.1
PyYAML==5.3.1

# analysis
matplotlib==3.3.0
networkx==2.4
numpy==1.19.0
pandas==1.0.5
scikit-learn==0.23.1
pytorch==1.5.1
torch-cluster==1.5.6
torch-geometric==1.6.0
torch-scatter==2.0.5
torch-sparse==0.6.6
torch-spline-conv==1.2.0
